clean and refactor by review
