+++ app/src/main/java/com/a1ishka/bankaccount/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.a1ishka.bankaccount.presentation.navigation.NavGraph
+import com.a1ishka.bankaccount.navigation.NavGraph
 import com.a1ishka.bankaccount.ui.theme.BankAccountTheme
 import dagger.hilt.android.AndroidEntryPoint
