+++ app/src/main/java/com/a1ishka/bankaccount/presentation/account/AccountEvent.kt
@@ -1,10 +1,10 @@
 package com.a1ishka.bankaccount.presentation.account
 
-import com.a1ishka.bankaccount.data.entity.AccountEntity
+import com.a1ishka.bankaccount.domain.Account
 
 sealed interface AccountEvent {
     object SaveAccount : AccountEvent
-    data class DeleteAccount(val account: AccountEntity) : AccountEvent
+    data class DeleteAccount(val account: Account) : AccountEvent
     data class SetAccountName(val name: String) : AccountEvent
     data class SetCardNumber(val cardNumber: Long) : AccountEvent
     data class SetAccountNumber(val accountNumber: Long) : AccountEvent
