+++ app/src/main/java/com/a1ishka/bankaccount/presentation/account/AccountState.kt
@@ -1,10 +1,10 @@
 package com.a1ishka.bankaccount.presentation.account
 
-import com.a1ishka.bankaccount.data.entity.AccountEntity
+import com.a1ishka.bankaccount.domain.Account
 
 data class AccountState(
-    val accountList: List<AccountEntity> = emptyList(),
-    val currentAccount: AccountEntity? = null,
+    val accountList: List<Account> = emptyList(),
+    val currentAccount: Account? = null,
     val name: String = "",
     val cardNumber: Long = 0,
     val accountNumber: Long = 0,
