+++ app/src/main/java/com/a1ishka/bankaccount/presentation/screens/accountDashboard/AccountDashboardScreen.kt
@@ -29,7 +29,7 @@ import androidx.navigation.NavController
 import com.a1ishka.bankaccount.R
 import com.a1ishka.bankaccount.data.accountData
 import com.a1ishka.bankaccount.data.transactionData
-import com.a1ishka.bankaccount.presentation.navigation.Screen
+import com.a1ishka.bankaccount.navigation.Screen
 
 @OptIn(ExperimentalMaterial3Api::class)
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
