+++ app/src/main/java/com/a1ishka/bankaccount/presentation/transaction/TransactionEvent.kt
@@ -1,13 +1,13 @@
 package com.a1ishka.bankaccount.presentation.transaction
 
-import com.a1ishka.bankaccount.data.entity.TransactionEntity
+import com.a1ishka.bankaccount.domain.Transaction
 
 sealed interface TransactionEvent {
     object SaveTransaction : TransactionEvent
     data class FilterTransactions(val accountId: Long, val startDate: String, val endDate: String) :
         TransactionEvent
 
-    data class DeleteTransaction(val transaction: TransactionEntity) : TransactionEvent
+    data class DeleteTransaction(val transaction: Transaction) : TransactionEvent
     data class SetApplier(val applier: String) : TransactionEvent
     data class SetNumber(val number: String) : TransactionEvent
     data class SetDate(val date: String) : TransactionEvent
